refactor(recording): move gesture telemetry out of the daemon zone
`recording-telemetry.ts` imports only `node:fs`, `node:path` and
`@agent-device/contracts/platform` — it has no daemon dependency and is
recording-domain file serialization, not daemon behavior. Sitting under
`src/daemon/` made the platform-runtime composition root reach up into
daemon-server (rank 4) for it.

Move it to `src/recording/telemetry.ts`, alongside `output-path.ts` and
`overlay.ts`, which share its shape (node builtins plus contracts/kernel).
Pure rename: the module body is unchanged and the three import sites are
retargeted.

This removes one of the composition root's four upward value edges — the
only one reaching rank 4. Confirmed with `pnpm depgraph`: those edges drop
from 4 to 3 and the daemon-server target is gone, with cycle, back-edge and
type-inversion counts unchanged.
