+++ src/recording/telemetry.ts
